dash/requirements.txt
dash==2.14.2
dash-auth==2.0.0
clickhouse-connect==0.6.22
pandas==2.1.4
gunicorn==21.2.0
python-dotenv==1.0.0
plotly